Split CSV document loading helpers
